feat(power up/down): check if a v.balance can be withdrawn
this check is based on whether or not the vesting_cliff_seconds + begin_timestamp of a vesting
balance exceeds the current time. If a gpos vested balance dow not meeth this condition, it cannot
be withdrawn/claimed. Results in potential for zero gpos vested balances to be withdrawn/claimed
despite the UI indicating otherwise.

WAL-250