v1.5.3-alpha hotfixes (#126)
* fix(done.jsx): participation Dialogue Not Displayed after voting->done

After voting, "Thank you for Voting" dialog will be open. Click on the OK button will now send user
to start.jsx

GPOS-52 #86

* refactor: remove dead code

* fix: updated react-tabs

there were issues with the current severely outdated version so the package has been updated. The
codebase has had some tweaks to account for this change: in the context of VotingContainer.jsx the
code has been simplified.

GPOS-50, #90

* fix: typos in some headers on vote witness/advisor

re-evaluated the helper render function to account for more variety of conditional texts

GPOS-64

* fix: fix maths again for power up/down

first gets taken from the vested gpos amount being withdrawn

GPOS-43

* fix: fix spacing on currency symbol on advisors tables

* fix(voting): vanishing prior votes

when a user votes for a witness and has broadcasted/published this, and then votes for an advisor
and has broadcasted/published this before the last broadcast is received in a new block on the
frontend: the first broadcast/publish is nullified but the user is still charged. SO, we add a
cooldown to the "done" button on transaction confirm modal to ensure a new block has been generated
and received by the frontend thereby making it impossible to publish another vote transaction
between block generation/updates

GPOS-61 #91, GPOS-50 #90

* fix(codesmells): resolve sonarcloud codesmells